+++ src/systems/si/units/mod.rs
@@ -8,9 +8,10 @@ pub use base::{Ampere, Candela, Kelvin, Kilogram, Meter, Mole, Second};
 pub use derived::{
     CubicMeter, CubicMeterPerSecond, Gray, Hertz, Joule, JoulePerCubicMeter, JoulePerKelvin,
     JoulePerKilogramKelvin, JoulePerMole, JoulePerMoleKelvin, JoulePerSquareMeter,
-    KilogramPerCubicMeter, KilogramPerCubicMeterSecond, MeterPerSecond, Pascal, PascalSecond,
-    PerKelvin, PerMeter, PerSecond, PerSquareKelvin, Rayl, SquareMeter, SquareMeterPerKilogram,
-    SquareMeterPerSecond, Watt, WattPerCubicMeter, WattPerMeterKelvin, WattPerSquareMeter,
+    KilogramPerCubicMeter, KilogramPerCubicMeterSecond, MeterPerSecond, NewtonPerMeter, Pascal,
+    PascalSecond, PerKelvin, PerMeter, PerSecond, PerSquareKelvin, Rayl, SquareMeter,
+    SquareMeterPerKilogram, SquareMeterPerSecond, Watt, WattPerCubicMeter, WattPerMeterKelvin,
+    WattPerSquareMeter,
 };
 pub use scaled::{
     Centimeter, CubicMillimeter, ElectronVolt, Gram, GramPerCubicCentimeter, JoulePerMilliliter,